test: move tests from postgres.rs to response_cache.rs
