+++ apollo-router/tests/integration/mod.rs
@@ -26,9 +26,6 @@ mod typename;
 
 // In the CI environment we only install PostgreSQL on x86_64 Linux
 #[cfg(any(not(feature = "ci"), all(target_arch = "x86_64", target_os = "linux")))]
-mod postgres;
-// In the CI environment we only install PostgreSQL on x86_64 Linux
-#[cfg(any(not(feature = "ci"), all(target_arch = "x86_64", target_os = "linux")))]
 mod response_cache;
 // In the CI environment we only install Redis on x86_64 Linux
 #[cfg(any(not(feature = "ci"), all(target_arch = "x86_64", target_os = "linux")))]
